[241226] Ship 클래스. 스킬 추가.
abstract를 이용하여 각 함급에 스킬을 추가
