+++ FleetComander_Console.csproj
@@ -44,13 +44,14 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ConsoleUiSetting.cs" />
+    <Compile Include="BattleSetting_Part2.cs" />
     <Compile Include="GameInfo.cs" />
     <Compile Include="MainMenu.cs" />
     <Compile Include="PlaySequence.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ShopSetting.cs" />
-    <Compile Include="BattleSetting.cs" />
+    <Compile Include="BattleSetting_Part1.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
